pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "beautifulsoup4>=4.14.3",
    "docling>=2.92.0",
    "fastapi-limiter>=0.2.0",
    "fastapi[standard]>=0.136.1",
    "fastembed>=0.8.0",
    "gunicorn>=25.3.0",
    "httpx>=0.28.1",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.2",
    "langchain-openai>=1.2.1",
    "langchain-redis>=0.2.5",
    "langchain-text-splitters>=1.1.2",
    "loguru>=0.7.3",
    "passlib>=1.7.4",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "qdrant-client>=1.17.1",
    "sentence-transformers>=5.4.1",
    "sqlalchemy>=2.0.49",
    "trafilatura>=2.0.0",
]
